changed Parameter from int to double to diplay the aimRAtio in a sensefull way

+++ src/main/java/org/visab/newgui/visualize/cbrshooter/model/CBRShooterImplicator.java
@@ -27,7 +27,7 @@ public final class CBRShooterImplicator {
         for (int i = 0; i < file.getStatistics().size(); i++) {
 
             if (round < file.getStatistics().get(i).getRound()) {
-                shotsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, countShots));
+                shotsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, (double) countShots));
                 countShots = 0;
             }
             maxAmmunition = currentAmmunition;
@@ -56,7 +56,8 @@ public final class CBRShooterImplicator {
         for (int i = 0; i < file.getStatistics().size(); i++) {
 
             if (round < file.getStatistics().get(i).getRound()) {
-                collectedCollectablesPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, countCollectedCollectables));
+                collectedCollectablesPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, 
+                        (double) countCollectedCollectables));
             }
             
             switch (collectable) {
@@ -104,9 +105,9 @@ public final class CBRShooterImplicator {
 
             if (shotsFired.get(i).getParameter() > 0) {
                 aimRatioPerRoundPerPlayer.add(new StatisticsDataStructure((double) i, 
-                        (hits.get(i).getParameter() / shotsFired.get(i).getParameter())));
+                        (double) (hits.get(i).getParameter() / shotsFired.get(i).getParameter())));
             } else {
-                aimRatioPerRoundPerPlayer.add(new StatisticsDataStructure((double) i, 0));
+                aimRatioPerRoundPerPlayer.add(new StatisticsDataStructure((double) i, 0.0));
             }
                 
         }
@@ -128,7 +129,7 @@ public final class CBRShooterImplicator {
         for (int i = 0; i < file.getStatistics().size(); i++) {
 
             if (round < file.getStatistics().get(i).getRound()) {
-                hitsTakenPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, countHits));
+                hitsTakenPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, (double) countHits));
                 countHits = 0;
             }
             lastHealth = currentHealth;
@@ -159,7 +160,7 @@ public final class CBRShooterImplicator {
         for (int i = 0; i < file.getStatistics().size(); i++) {
 
             if (round < file.getStatistics().get(i).getRound()) {
-                accumulatedDeathsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, unitsWalked));
+                accumulatedDeathsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, (double) unitsWalked));
                 unitsWalked = 0;
             }
             
@@ -190,7 +191,7 @@ public final class CBRShooterImplicator {
 
             if (round < file.getStatistics().get(i).getRound()) {
                 accumulatedDeathsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, 
-                        file.getStatistics().get(i).getPlayers().get(playerNumber).getStatistics().getDeaths()));
+                        (double) file.getStatistics().get(i).getPlayers().get(playerNumber).getStatistics().getDeaths()));
             }
 
             round = file.getStatistics().get(i).getRound();
@@ -211,7 +212,7 @@ public final class CBRShooterImplicator {
 
             if (round < file.getStatistics().get(i).getRound()) {
                 accumulatedKillsPerRoundPerPlayer.add(new StatisticsDataStructure((double) round, 
-                        file.getStatistics().get(i).getPlayers().get(playerNumber).getStatistics().getDeaths()));
+                        (double) file.getStatistics().get(i).getPlayers().get(playerNumber).getStatistics().getDeaths()));
             }
 
             round = file.getStatistics().get(i).getRound();
